refactor: Extract into `TaxonDecorator::HeadlineType`
